+++ src/context/index.js
@@ -1,0 +1,5 @@
+import { FilterProvider } from "./Filtercontext";
+import { useFilter } from "./Filtercontext";
+
+
+export { FilterProvider, useFilter};
