+++ src/index.js
@@ -4,7 +4,7 @@ import "./index.css";
 import App from "./App";
 import { makeServer } from "./server";
 import { BrowserRouter } from "react-router-dom";
-import { FilterProvider } from "./context/Filtercontext";
+import { FilterProvider } from "./context";
 
 // Call make Server
 makeServer();
